+++ jablib/src/main/java/org/jabref/model/entry/identifier/Identifier.java
@@ -2,8 +2,11 @@ package org.jabref.model.entry.identifier;
 
 import java.net.URI;
 import java.util.Optional;
+import java.util.function.Supplier;
+import java.util.stream.Stream;
 
 import org.jabref.model.entry.field.Field;
+import org.jabref.model.strings.StringUtil;
 
 /**
  * All implementing classes should additionally offer
@@ -23,4 +26,23 @@ public interface Identifier {
     Field getDefaultField();
 
     Optional<URI> getExternalURI();
+
+    public static Optional<Identifier> from(String identifier) {
+        if (StringUtil.isBlank(identifier)) {
+            return Optional.empty();
+        }
+
+        return Stream.<Supplier<Optional<? extends Identifier>>>of(
+                             () -> DOI.findInText(identifier),
+                             () -> ArXivIdentifier.parse(identifier),
+                             () -> ISBN.parse(identifier),
+                             () -> SSRN.parse(identifier),
+                             () -> RFC.parse(identifier)
+                     )
+                     .map(Supplier::get)
+                     .filter(Optional::isPresent)
+                     .map(Optional::get)
+                     .map(id -> (Identifier) id)
+                     .findFirst();
+    }
 }