+++ jablib/src/test/java/org/jabref/logic/quality/consistency/BibliographyConsistencyCheckResultTxtWriterTest.java
@@ -25,7 +25,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 
 class BibliographyConsistencyCheckResultTxtWriterTest {
-    private BibtexImporter importer = new BibtexImporter(mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS), new DummyFileUpdateMonitor());
+    private final BibtexImporter importer = new BibtexImporter(mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS), new DummyFileUpdateMonitor());
 
     @Test
     void checkSimpleLibrary(@TempDir Path tempDir) throws IOException {
@@ -35,7 +35,7 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
         BibEntry second = new BibEntry(StandardEntryType.Article, "second")
                 .withField(StandardField.AUTHOR, "Author One")
                 .withField(StandardField.PUBLISHER, "publisher");
-        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second));
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second), (_, _) -> { });
 
         Path txtFile = tempDir.resolve("checkSimpleLibrary-result.txt");
         try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
@@ -50,10 +50,12 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
                 | Article    | first        | o     | -         |
                 | Article    | second       | -     | ?         |
 
-                x | required field is present
-                o | optional field is present
-                ? | unknown field is present
-                - | field is absent
+                | Symbol | Meaning                   |
+                | ------ | ------------------------- |
+                | x      | required field is present |
+                | o      | optional field is present |
+                | ?      | unknown field is present  |
+                | -      | field is absent           |
                 """, Files.readString(txtFile).replace("\r\n", "\n"));
     }
 
@@ -67,7 +69,7 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
                 .withField(customField, "custom"); // unknown
         BibEntry second = new BibEntry(StandardEntryType.Article, "second")
                 .withField(StandardField.AUTHOR, "Author One");
-        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second));
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second), (_, _) -> { });
 
         Path txtFile = tempDir.resolve("checkDifferentOutputSymbols-result.txt");
         try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
@@ -81,10 +83,45 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
                 | ---------- | ------------ | ------ | ----- | ----- |
                 | Article    | first        | ?      | o     | x     |
 
-                x | required field is present
-                o | optional field is present
-                ? | unknown field is present
-                - | field is absent
+                | Symbol | Meaning                   |
+                | ------ | ------------------------- |
+                | x      | required field is present |
+                | o      | optional field is present |
+                | ?      | unknown field is present  |
+                | -      | field is absent           |
+                """, Files.readString(txtFile).replace("\r\n", "\n"));
+    }
+
+    @Test
+    void checkVeryLongCitationKey(@TempDir Path tempDir) throws IOException {
+        UnknownField customField = new UnknownField("custom");
+        BibEntry first = new BibEntry(StandardEntryType.Article, "first-very-long-key")
+                .withField(StandardField.AUTHOR, "Author One") // required
+                .withField(StandardField.TITLE, "Title") // required
+                .withField(StandardField.PAGES, "some pages") // optional
+                .withField(customField, "custom"); // unknown
+        BibEntry second = new BibEntry(StandardEntryType.Article, "second")
+                .withField(StandardField.AUTHOR, "Author One");
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second), (_, _) -> { });
+
+        Path txtFile = tempDir.resolve("checkDifferentOutputSymbols-result.txt");
+        try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
+             BibliographyConsistencyCheckResultTxtWriter BibliographyConsistencyCheckResultTxtWriter = new BibliographyConsistencyCheckResultTxtWriter(result, writer, false)) {
+            BibliographyConsistencyCheckResultTxtWriter.writeFindings();
+        }
+        assertEquals("""
+                Field Presence Consistency Check Result
+
+                | entry type | citation key        | Custom | Pages | Title |
+                | ---------- | ------------------- | ------ | ----- | ----- |
+                | Article    | first-very-long-key | ?      | o     | x     |
+
+                | Symbol | Meaning                   |
+                | ------ | ------------------------- |
+                | x      | required field is present |
+                | o      | optional field is present |
+                | ?      | unknown field is present  |
+                | -      | field is absent           |
                 """, Files.readString(txtFile).replace("\r\n", "\n"));
     }
 
@@ -106,11 +143,16 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
                 .withField(StandardField.AUTHOR, "Author One")
                 .withField(StandardField.YEAR, "2024")
                 .withField(StandardField.PUBLISHER, "publisher");
+        // Entry added to check for alphabetical ordering of citation keys
         BibEntry fifth = new BibEntry(StandardEntryType.InProceedings, "fifth")
+                .withField(StandardField.AUTHOR, "Author One")
+                .withField(StandardField.LOCATION, "location")
+                .withField(StandardField.YEAR, "2024");
+        BibEntry sixth = new BibEntry(StandardEntryType.InProceedings, "sixth")
                 .withField(StandardField.AUTHOR, "Author One")
                 .withField(StandardField.YEAR, "2024");
 
-        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second, third, fourth, fifth));
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second, third, fourth, fifth, sixth), (_, _) -> { });
 
         Path txtFile = tempDir.resolve("checkSimpleLibrary-result.txt");
         try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
@@ -120,17 +162,20 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
         assertEquals("""
                 Field Presence Consistency Check Result
 
-                | entry type    | citation key  | Location | Pages | Publisher |
-                | ------------- | ------------- | -------- | ----- | --------- |
-                | Article       | first         | -        | o     | -         |
-                | Article       | second        | -        | -     | ?         |
-                | InProceedings | fourth        | -        | -     | o         |
-                | InProceedings | third         | ?        | o     | -         |
+                | entry type    | citation key | Location | Pages | Publisher |
+                | ------------- | ------------ | -------- | ----- | --------- |
+                | Article       | first        | -        | o     | -         |
+                | Article       | second       | -        | -     | ?         |
+                | InProceedings | fifth        | ?        | -     | -         |
+                | InProceedings | fourth       | -        | -     | o         |
+                | InProceedings | third        | ?        | o     | -         |
 
-                x | required field is present
-                o | optional field is present
-                ? | unknown field is present
-                - | field is absent
+                | Symbol | Meaning                   |
+                | ------ | ------------------------- |
+                | x      | required field is present |
+                | o      | optional field is present |
+                | ?      | unknown field is present  |
+                | -      | field is absent           |
                 """, Files.readString(txtFile).replace("\r\n", "\n"));
     }
 
@@ -142,7 +187,7 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
         BibEntry second = new BibEntry(StandardEntryType.Article, "second")
                 .withField(StandardField.AUTHOR, "Author One")
                 .withField(StandardField.PAGES, "some pages");
-        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second));
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second), (_, _) -> { });
 
         Path txtFile = tempDir.resolve("checkLibraryWithoutIssues-result.txt");
         try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
@@ -164,7 +209,7 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
         BibEntry second = new BibEntry(StandardEntryType.Article, "second")
                 .withField(StandardField.AUTHOR, "Author One")
                 .withField(StandardField.PAGES, "some pages");
-        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second));
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(List.of(first, second), (_, _) -> { });
 
         Path txtFile = tempDir.resolve("checkLibraryWithoutIssues-result.txt");
         try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
@@ -180,7 +225,7 @@ class BibliographyConsistencyCheckResultTxtWriterTest {
         Path file = Path.of("C:\\TEMP\\JabRef\\biblio-anon.bib");
         Path txtFile = file.resolveSibling("biblio-cited.txt");
         BibDatabaseContext databaseContext = importer.importDatabase(file).getDatabaseContext();
-        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(databaseContext.getEntries());
+        BibliographyConsistencyCheck.Result result = new BibliographyConsistencyCheck().check(databaseContext.getEntries(), (_, _) -> { });
         try (Writer writer = new OutputStreamWriter(Files.newOutputStream(txtFile));
              BibliographyConsistencyCheckResultTxtWriter BibliographyConsistencyCheckResultTxtWriter = new BibliographyConsistencyCheckResultTxtWriter(result, writer, true)) {
             BibliographyConsistencyCheckResultTxtWriter.writeFindings();