+++ jabsrv/src/main/java/org/jabref/http/server/LibraryResource.java
@@ -1,6 +1,10 @@
 package org.jabref.http.server;
 
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.List;
 import java.util.Objects;
@@ -13,6 +17,7 @@ import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.importer.fileformat.BibtexImporter;
 import org.jabref.logic.preferences.CliPreferences;
 import org.jabref.logic.util.io.BackupFileUtil;
+import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.util.DummyFileUpdateMonitor;
 
@@ -27,6 +32,8 @@ import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.StreamingOutput;
+import org.jspecify.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -45,7 +52,7 @@ public class LibraryResource {
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public String getJson(@PathParam("id") String id) {
+    public String getJson(@PathParam("id") String id) throws IOException {
         ParserResult parserResult = getParserResult(id);
         BibEntryTypesManager entryTypesManager = Injector.instantiateModelOrService(BibEntryTypesManager.class);
         List<BibEntryDTO> list = parserResult.getDatabase().getEntries().stream()
@@ -57,28 +64,29 @@ public class LibraryResource {
 
     @GET
     @Produces(JabrefMediaType.JSON_CSL_ITEM)
-    public String getClsItemJson(@PathParam("id") String id) {
+    public String getClsItemJson(@PathParam("id") String id) throws IOException {
         ParserResult parserResult = getParserResult(id);
         JabRefItemDataProvider jabRefItemDataProvider = new JabRefItemDataProvider();
         jabRefItemDataProvider.setData(parserResult.getDatabaseContext(), new BibEntryTypesManager());
         return jabRefItemDataProvider.toJson();
     }
 
-    private ParserResult getParserResult(String id) {
-        java.nio.file.Path library = getLibraryPath(id);
-        ParserResult parserResult;
-        try {
-            parserResult = new BibtexImporter(preferences.getImportFormatPreferences(), new DummyFileUpdateMonitor()).importDatabase(library);
-        } catch (IOException e) {
-            LOGGER.warn("Could not find open library file {}", library, e);
-            throw new InternalServerErrorException("Could not parse library", e);
-        }
-        return parserResult;
-    }
-
     @GET
     @Produces(JabrefMediaType.BIBTEX)
     public Response getBibtex(@PathParam("id") String id) {
+        if ("demo".equals(id)) {
+            StreamingOutput stream = output -> {
+                try (InputStream in = getChocolateBibAsStream()) {
+                    in.transferTo(output);
+                }
+            };
+
+            return Response.ok(stream)
+                           // org.glassfish.jersey.media would be required for a "nice" Java to create ContentDisposition; we avoid this
+                           .header("Content-Disposition", "attachment; filename=\"Chocolate.bib\"")
+                           .build();
+        }
+
         java.nio.file.Path library = getLibraryPath(id);
         String libraryAsString;
         try {
@@ -88,6 +96,7 @@ public class LibraryResource {
             throw new InternalServerErrorException("Could not read library " + library, e);
         }
         return Response.ok()
+                .header("Content-Disposition", "attachment; filename=\"" + library.getFileName() + "\"")
                 .entity(libraryAsString)
                 .build();
     }
@@ -98,5 +107,31 @@ public class LibraryResource {
                           .filter(p -> (p.getFileName() + "-" + BackupFileUtil.getUniqueFilePrefix(p)).equals(id))
                           .findAny()
                           .orElseThrow(NotFoundException::new);
+    }
+
+    private ParserResult getParserResult(String id) throws IOException {
+        BibtexImporter bibtexImporter = new BibtexImporter(preferences.getImportFormatPreferences(), new DummyFileUpdateMonitor());
+
+        if ("demo".equals(id)) {
+            try (InputStream chocolateBibInputStream = getChocolateBibAsStream()) {
+                BufferedReader reader = new BufferedReader(new InputStreamReader(chocolateBibInputStream, StandardCharsets.UTF_8));
+                return bibtexImporter.importDatabase(reader);
+            }
+        }
+
+        java.nio.file.Path library = getLibraryPath(id);
+        ParserResult parserResult;
+        try {
+            parserResult = bibtexImporter.importDatabase(library);
+        } catch (IOException e) {
+            LOGGER.warn("Could not find open library file {}", library, e);
+            throw new InternalServerErrorException("Could not parse library", e);
+        }
+        return parserResult;
+    }
+
+    /// @return a stream to the Chocolate.bib file in the classpath (is null only if the file was moved or there are issues with the classpath)
+    private @Nullable InputStream getChocolateBibAsStream() {
+        return BibDatabase.class.getResourceAsStream("/Chocolate.bib");
     }
 }