+++ jabsrv/src/test/java/org/jabref/http/server/LibrariesResourceTest.java
@@ -23,8 +23,9 @@ class LibrariesResourceTest extends ServerTest {
     void defaultOneTestLibrary() {
         String expected = """
                 [
+                  "%s",
                   "%s"
-                ]""".formatted(TestBibFile.GENERAL_SERVER_TEST.id);
+                ]""".formatted(TestBibFile.GENERAL_SERVER_TEST.id, "demo");
         assertEquals(expected, target("/libraries").request().get(String.class));
     }
 
@@ -36,8 +37,9 @@ class LibrariesResourceTest extends ServerTest {
         String expected = """
                 [
                   "%s",
+                  "%s",
                   "%s"
-                ]""".formatted(TestBibFile.GENERAL_SERVER_TEST.id, TestBibFile.CHOCOLATE_BIB.id);
+                ]""".formatted(TestBibFile.GENERAL_SERVER_TEST.id, TestBibFile.CHOCOLATE_BIB.id, "demo");
         assertEquals(expected, target("/libraries").request().get(String.class));
     }
 }